requirements.txt
flask
flask_bcrypt
flask_cors
flask_sqlalchemy
flask_login
flask_jwt
flask_jwt_extended
flask_mail
flask_restful
flask_testing 
jsonify
psycopg2

